requirements.txt
flask
fastapi
pyngrok
kafka-python
pyyaml
pydantic
langgraph
langchain
requests
pytest
uvicorn
pytest-shutil
httpx
subprocess.run